fix(pdf): Date.UTC years 0-99 must use setUTCFullYear
Extract lightweight pdf-date so regression test imports production without @noble/hashes heavy import.
